requirements.txt
Appium-Python-Client==0.47
pycryptodome==3.8.2
schedule==0.6.0
acm-sdk-python==0.4.11
altgraph==0.15
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.7.1
billiard==3.6.0.0
bleach==3.1.0
bs4==0.0.1
celery==4.3.0
celery_once==3.0.1
certifi==2019.6.16
cffi==1.11.4
chardet==3.0.4
click==6.7
colorama==0.4.1
common==0.1.2
config==0.3.9
coverage==4.5.3
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dnspython==1.16.0
docopt==0.6.2
elasticsearch==7.0.2
entrypoints==0.3
et-xmlfile==1.0.1
eventlet==0.25.0
fishbase==1.1.14
flake8==3.7.7
Flask==0.12.2
future==0.16.0
greenlet==0.4.15
idna==2.8
importlib-metadata==0.18
inflect==2.1.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==0.24
jdcal==1.4.1
jedi==0.14.0
Jinja2==2.10
jinja2-pluralize==0.3.0
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.0
jupyter-console==6.0.0
jupyter-core==4.5.0
kazoo==2.5.0
kiwisolver==1.1.0
kombu==4.6.3
lazy-object-proxy==1.4.1
macholib==1.9
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==7.0.0
nbconvert==5.5.0
nbformat==4.4.0
nose==1.3.0
notebook==5.7.8
numpy==1.16.4
openpyxl==2.6.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parameterized==0.6.1
paramiko==2.4.2
parso==0.5.0
pefile==2017.11.5
pickleshare==0.7.5
pluggy==0.12.0
portalocker==0.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
py==1.8.0
pyasn1==0.4.4
pycodestyle==2.5.0
pycparser==2.18
pydocstyle==3.0.0
pyflakes==2.1.1
Pygments==2.4.2
PyInstaller==3.3.1
pylint==2.3.1
pymongo==3.7.2
PyMySQL==0.8.0
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyparsing==2.4.0
pypiwin32==223
pyrsistent==0.15.3
PySocks==1.6.7
pytest==4.6.3
pytest-cov==2.7.1
python-dateutil==2.8.0
pytz==2019.1
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
redis==3.3.5
requests==2.22.0
rstr==2.2.6
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9.2
SQLAlchemy==1.2.8
stomp.py==4.1.21
terminado==0.8.2
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==0.5.1
flower==0.9.3
